refactor(cashier): create SalesCard.vue component
move _Current sale_ card to their own component from Monolith.vue
